Refactored Summary files, updated util.js
